+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v1.1.0] - 2025-03-05
+
+### Fixes
+
+- Shop keepers in Gold Rock City, Sage Hills, Eastwind, and Chronos not following the sundown schedule
+- Not being able to buy anything in the Sage Hills inn at night
+- Guy in Happy Bay near the top of the town not following a schedule
+
 ## [v1.0.1] - 2024-09-10
 
 ### Fixes
@@ -13,9 +21,6 @@ All notable changes to this project will be documented in this file.
 ### Fixes
 
 - UI sounds playing when you are in another application
-
 - Fishing rod reel rotating the opposite way
-
 - Backup saves for save slot one not popping up when hovering over it until you hover over another save slot
-
 - Tree in autosave section of Settings menu